requirements.txt
streamlit==1.31.1
plotly==5.19.0
numpy>=1.24.0
pandas>=2.0.0
openpyxl>=3.1.2  # Para exportación a Excel
fpdf2>=2.7.8  # Para exportación a PDF
kaleido>=0.2.1  # Para guardar gráficos de Plotly
